Add ContentTabsMenu to mobile
