+++ src/features/content/components/content-panel.tsx
@@ -1,31 +1,11 @@
 import { Link } from "@tanstack/react-router"
-import { styled } from "@linaria/react"
-import { Flex, IconButton, Spacer } from "@chakra-ui/react"
-import { FiChevronLeft, FiChevronRight } from "react-icons/fi"
+import { Spacer } from "@chakra-ui/react"
 import Panel, { PanelCloseButton, PanelContent } from "@/components/panel"
-import type { ContentProps } from "./content-container"
+import { ContentText, type ContentProps } from "./content-container"
 import Loadable from "@/components/loadable"
-import ContentTabsMenu, { handleTabChange } from "./content-tabs-menu"
-
-const ContentText = styled.div`
-  color: black;
-  font-family: Bitter, sans-serif;
-  font-weight: 400;
-  padding: 8px;
-  padding-top: 0.5rem;
-  display: flex;
-  flex-direction: column;
-`
+import ContentTabsMenu, { ContentTabsPrevNextFooter } from "./content-tabs-menu"
 
 export default function ContentPanel(props: ContentProps) {
-  // Component assumed to be isMobileOrTablet = false
-  const activeTabIndex = props.tabs.findIndex((tab) => tab.id === props.activeTabId)
-  const prevTab = activeTabIndex-1 < 0 || activeTabIndex-1 >= props.tabs.length
-    ? null
-    : props.tabs[activeTabIndex-1].id
-  const nextTab = activeTabIndex+1 < 0 || activeTabIndex+1 >= props.tabs.length
-    ? null
-    : props.tabs[activeTabIndex+1].id
 
   return (
     <Panel
@@ -45,41 +25,10 @@ export default function ContentPanel(props: ContentProps) {
           <ContentText>
             { props.children }
             <Spacer />
-            <Flex
-              direction='row'
-              justifyContent='space-between'
-              paddingTop='0.5rem'
-              paddingBottom='2px'
-            >
-              <PrevNextButton direction="prev" tabId={prevTab} />
-              <PrevNextButton direction="next" tabId={nextTab} />
-            </Flex>
+            <ContentTabsPrevNextFooter tabs={props.tabs} activeTabId={props.activeTabId} />
           </ContentText>
         </Loadable>
       </PanelContent>
     </Panel>
-  )
-}
-
-function PrevNextButton(props: { direction: 'prev' | 'next', tabId: string | null }) {
-  const ariaLabel = {
-    prev: 'Previous tab',
-    next: 'Next tab'
-  }
-  return (
-    props.tabId 
-      ? (
-        <Link to='/app/$topicId/$tabId' params={{tabId: props.tabId}}>
-          <IconButton
-            onClick={() => handleTabChange(props.tabId)}
-            aria-label={ariaLabel[props.direction]}
-            variant='surface'
-          >
-            {props.direction === 'prev' ? <FiChevronLeft /> : <FiChevronRight />}
-          </IconButton>
-        </Link>
-      ) : (
-        <Spacer />
-      )
   )
 }
